backend/requirements.txt
# FastAPI & ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Django
django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# ML Dependencies
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# LightGBM - use pip install with no build (pre-built binary)
# Install separately: pip install lightgbm --no-build-isolation
lightgbm

# Utils
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2

# CORS & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
